+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Write metadata even if `git` can't be found on `PATH`. 
+
 ## [0.1.0.dev11] - 2021-02-28
 ### Added
 - [#5] *unittest* integration.